merge IEC104 Branch with Development Branch

* commit '320e5c706ced75138a73a5b27eb345bbcec831d3':
  [General] Minor Adjustment
  [Feature] Report Feature Complete.
  [Feature] Report Feature Complete.
  [General] Controller & Script Cleanup
  [Feature] DO Report Function
  [General] Modify Database
  [General] Implemented DO2 Test Case. Invalid negative response
  [General] InitIEC104OutputConnection
  [General] Output Hex Digits Success
  [General] Add DO Receiver for IEC-104 Command. Print console messages to debuglog

+++ src/script/TriggerModbusPointDI2.java
@@ -53,7 +53,7 @@ public class TriggerModbusPointDI2 implements InterfaceScript {
      */
     public void updateHoldingRegister(SimpleProcessImage processImage, int register, double newValue) throws InterruptedException {
 
-        processImage.setRegister((int) register, new SimpleRegister((int) newValue));
+        processImage.setRegister(register, new SimpleRegister((int) newValue));
 
         //CommonFunctions.debugLog.error("Register is : " + newValue + " value is :" + register);
         Thread.sleep(500);
@@ -61,7 +61,7 @@ public class TriggerModbusPointDI2 implements InterfaceScript {
 
     private void updateInputRegister(SimpleProcessImage instance, int register, double value) throws InterruptedException {
 
-        instance.setInputRegister((int) register, new SimpleRegister((int) value));
+        instance.setInputRegister(register, new SimpleRegister((int) value));
         Thread.sleep(500);
     }
